Replace FakeClient with envtest
As stated at controller-runtime docs ([1], [2]) FakeClient is going to be deprecated at 1.0.0,
re-implementation of controller-runtime client that's is not being maintained and
missing features, the replacement is envtest.

The envtest do a etcd/kube-apiserver binaries start/stop and allow you to prepare
a real runtime-controller client or k8s clientset, allowing you to test the code
with a real client but without a full cluster.

Another interesting features is that with some flag setting same tests can be run
at a proper cluster.

[1] https://godoc.org/github.com/kubernetes-sigs/controller-runtime/pkg/client/fake
[2] https://github.com/kubernetes-sigs/controller-runtime/issues/768#issuecomment-642335552

Signed-off-by: Quique Llorente <ellorent@redhat.com>