Example and controller wrapper cleanup
The `test/pod` is no longer needed since `pkg/webhook/server/server_test.go` can
be used as an example since it already integrate this library into a controller-runtime
manager.

Also the controller wrapper to mix non leader aware controller with leader aware controller
is not needed for any project and also is out of scope of this library, is something like
that is needed the project has to implement it.

Signed-off-by: Quique Llorente <ellorent@redhat.com>