Implement CAOverlapInterval options to ditch old CAs from CABunlde
This commit add a knob to configure how much time time a CA certificate that is not
the latest generated on after rotation can live at
CABunlde instead of using expiration time, if it's not set it uses CARotateInterval
which is similar to expiration time. This reduce the size of CABundle.

Signed-off-by: Quique Llorente <ellorent@redhat.com>