Support multiple webhooks & refactor
The primary concern of this commit is to add support for handling
multiple webhooks with a single certificate manager/controller.

This lead to a refactor to isolate certificate handling from
configuration handling.

A new package certificate/chain defines data structures to represent a
certificate chain and to ensure that certificate chain is current in
regards to expiration and overlap times, rotating where necessary.

All configuration handling has been refactored in configuration.go with
a generic concept of mapping runtime objects to/from certificate chain.

Other functional changes include:

* Drop support for URL client config in webhooks. It does not make sense
to manage certificates for webhooks that are hosted outside of the
cluster.
* Removed retries. Assumed that the primary retry mechanism should be to
requeue a controller request versus handling retries internally.